chore: Cleanup guest pmem package
Move device mapper code from pmem to devicemapper package.
